requirements.txt
Django==1.11.2
googlemaps==2.5.0
line-bot-sdk==1.3.0
urllib3==1.21.1
virtualenv==15.1.0
gunicorn==19.7.1
pylibmc==1.5.1
django-pylibmc==0.6.1